+++ packages/expo-updates/android/src/main/java/expo/modules/updates/manifest/NewManifest.kt
@@ -53,7 +53,8 @@ class NewManifest private constructor(
       assetList.add(
         AssetEntity(
           mLaunchAsset.getString("key"),
-          mLaunchAsset.getString("contentType")
+          // the fileExtension is not necessary for the launch asset and EAS servers will not include it.
+          mLaunchAsset.optString("fileExtension")
         ).apply {
           url = Uri.parse(mLaunchAsset.getString("url"))
           isLaunchAsset = true
@@ -66,13 +67,15 @@ class NewManifest private constructor(
       for (i in 0 until mAssets.length()) {
         try {
           val assetObject = mAssets.getJSONObject(i)
-          assetList.add(AssetEntity(
-            assetObject.getString("key"),
-            assetObject.getString("contentType")
-          ).apply {
-            url = Uri.parse(assetObject.getString("url"))
-            embeddedAssetFilename = assetObject.optString("embeddedAssetFilename")
-          })
+          assetList.add(
+            AssetEntity(
+              assetObject.getString("key"),
+              assetObject.getString("fileExtension")
+            ).apply {
+              url = Uri.parse(assetObject.getString("url"))
+              embeddedAssetFilename = assetObject.optString("embeddedAssetFilename")
+            }
+          )
         } catch (e: JSONException) {
           Log.e(TAG, "Could not read asset from manifest", e)
         }